sqlite: Replace OFFSET with WHERE clauses based on index scans
OFFSET is very inefficient when looking for uncommon words, because the DBMS
needs to scan through potentially many records and discard them, just to count
the offset of the first record to return.

By using a WHERE claused based on indices like this one, it can skip directly
to the first candidate.